Who cares about diffs? not me
This was some hackyness that I added to make sure writes were actually
happening on the datastore.  Back in the day, we ran out of disk space a
lot and writes would never make it to disk.  This doesn't happen
anymore.

Monitoring the database queue size and exception rate should be enough.